requirements.txt
Django==1.10.2
django-bootstrap3==7.1.0
django-registration==2.1.2